+++ src/Response/Model/CameraModels.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace InstagramAPI\Response\Model;
+
+use InstagramAPI\AutoPropertyMapper;
+
+class CameraModels extends AutoPropertyMapper
+{
+    const JSON_PROPERTY_MAP = [
+    ];
+}